Adapt tests from tests/rcpsp_multiskills for pytest
