+++ tests_clean/rcpsp_multiskill/test_rcpsp_ms_ls.py
@@ -1,7 +1,11 @@
+import os
 from typing import Dict, List, Set
 
-from discrete_optimization.rcpsp.rcpsp_solvers import *
+# from tests.rcpsp_multiskills.solvers.instance_creator import create_ms_rcpsp_demo
 from discrete_optimization.rcpsp.solver.ls_solver import LS_SOLVER, LS_RCPSP_Solver
+from discrete_optimization.rcpsp_multiskill.plots.plot_solution import (
+    plot_resource_individual_gantt,
+)
 from discrete_optimization.rcpsp_multiskill.rcpsp_multiskill import (
     Employee,
     MS_RCPSPModel,
@@ -10,11 +14,10 @@ from discrete_optimization.rcpsp_multiskill.rcpsp_multiskill import (
     MS_RCPSPSolution_Variant,
     SkillDetail,
 )
-from discrete_optimization.rcpsp_multiskill.solvers.solver_rcpsp_based import (
-    Solver_RCPSP_Based,
+from discrete_optimization.rcpsp_multiskill.rcpsp_multiskill_parser import (
+    get_data_available,
+    parse_file,
 )
-
-from tests.rcpsp_multiskills.solvers.instance_creator import create_ms_rcpsp_demo
 
 
 def create_toy_msrcpsp():
@@ -88,25 +91,27 @@ def create_toy_msrcpsp():
     return model
 
 
-def run_rcpsp_based():
-    # model = create_toy_msrcpsp()
-    model, model_rcpsp = create_ms_rcpsp_demo()
+def test_ls():
+    model = create_toy_msrcpsp()
     model = model.to_variant_model()
-    print(solvers)
-    method = LP_MRCPSP
-    params = [
-        solvers[k][j][1]
-        for k in solvers
-        for j in range(len(solvers[k]))
-        if solvers[k][j][0] == method
-    ][0]
-    params["lp_solver"] = LP_RCPSP_Solver.GRB
-    solver = Solver_RCPSP_Based(model=model, method=method, **params)
-    result = solver.solve()
+    solver = LS_RCPSP_Solver(model=model, ls_solver=LS_SOLVER.SA)
+    result = solver.solve(nb_iteration_max=1000)
     solution: MS_RCPSPSolution = result.get_best_solution()
-    print("Evaluation ", model.evaluate(solution))
-    print("Satisfaction ", model.satisfy(solution))
+    model.evaluate(solution)
+    assert model.satisfy(solution)
+
+
+def test_ls_imopse():
+    file = [f for f in get_data_available() if "100_5_22_15.def" in f][0]
+    model, name_task = parse_file(file, max_horizon=1000)
+    model = model.to_variant_model()
+    solver = LS_RCPSP_Solver(model=model, ls_solver=LS_SOLVER.SA)
+    result = solver.solve(nb_iteration_max=1)
+    solution: MS_RCPSPSolution = result.get_best_solution()
+    model.evaluate(solution)
+    assert model.satisfy(solution)
+    plot_resource_individual_gantt(rcpsp_model=model, rcpsp_sol=solution)
 
 
 if __name__ == "__main__":
-    run_rcpsp_based()
+    test_ls_imopse()
